requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
authlib
python-jose
flask-cors
six
python-dotenv
requests
authlib